Added README for Next.js project
A new README file has been added to the project. This file provides instructions on how to get started with the development server, including different command options. It also introduces the use of `next/font` for font optimization and loading, and provides resources for learning more about Next.js. Finally, it includes details on deploying the app using Vercel platform.